Add support for OpenAI API
aiac will now use OpenAI's official (beta) API by default. This means we
can use the GPT-3 language mode, which returns code but not a Markdown
of explanations like ChatGPT does. The API uses a simple API key that
can be provided via the `--api-key` flag, or the `OPENAI_API_KEY`
environment variable.

The previous ChatGPT usage is still supported, but via the `--chat-gpt`
flag. When used, `--session-token` should also be provided (or the
`CHATGPT_SESSION_TOKEN` environment variable).

The application now moves its logic behind a command named "get" (alias
"generate"), so it's not a part of the prompt. Instead, the library will
always prepend the word "generate" to the prompt (if we simply used
"get", the model will not necessarily generate code).